+++ lib/last_auto_index.php
@@ -5,7 +5,7 @@
  * We are just going to import some presets to make things easier
  */
 
-$_lai = new stdClass();
+$_lai = new stdClass;
 
 // pre-functions
 require_once(__DIR__.'\pre-functions.php');
@@ -25,10 +25,10 @@ _define('SER_NAME'    , $_SERVER['SERVER_NAME']);
 _define('SER_HOST'    , $_SERVER['HTTP_HOST']);
 _define('SER_ADDRESS' , $_SERVER['SERVER_ADDR']);
 _define('SER_PORT'    , $_SERVER['SERVER_PORT']);
-_define('SER_FILENAME', $_SERVER['SCRIPT_NAME']);
-_define('SER_SELF'    , $_SERVER['PHP_SELF']);
-_define('SER_DOC_ROOT', $_SERVER['DOCUMENT_ROOT']);
-_define('SER_REQ_URI' , $_SERVER['REQUEST_URI']);
+_define('SER_FILENAME', urldecode($_SERVER['SCRIPT_NAME']));
+_define('SER_SELF'    , urldecode($_SERVER['PHP_SELF']));
+_define('SER_DOC_ROOT', urldecode($_SERVER['DOCUMENT_ROOT']));
+_define('SER_REQ_URI' , $_SERVER['REQUEST_URI']); // [comeback] urldecode path but not query
 _define('USER_AGENT'  , $_SERVER['HTTP_USER_AGENT']);
 _define('USER_QUERY'  , $_SERVER['QUERY_STRING']);
 _define('USER_IP'     , $_SERVER['REMOTE_ADDR']);