+++ plugins/file-viewer/file_viewer.class.php
@@ -2,6 +2,47 @@
 
 class file_viewer extends plugin {
 	
+	public function __construct(){
+		
+		
+		$this->init();
+	}
+	
+	private function init(){
+		
+	}
+	
+	public function set_working_file(){
+		
+		// return bool
+	}
+	
+	public function get_working_file(){
+		
+		// return str/false
+	}
+	
+	private function get_filetype(){
+		
+		// return str/false
+	}
+	
+	public function view(){
+		
+	}
+	
+	public function can_view(){
+		
+	}
+	
+	public function is_source(){
+		
+	}
+	
+	public function is_display(){
+		
+	}
+	
 	
 	
 }